+++ Masterthesis/Assets/Scripts/Noise/OctaveNoise/OctaveNoiseSettings.cs
@@ -16,8 +16,8 @@ namespace SBaier.Master
 		private float _startFrequency = 1;
 		public float StartFrequency => _startFrequency;
 		[SerializeField]
-		private double _startWeight = 1;
-		public double StartWeight => _startWeight;
+		private float _startWeight = 1;
+		public float StartWeight => _startWeight;
 
 		public override NoiseType GetNoiseType()
 		{
